[sw/device/sca] Prepare to add C warning flags

+++ sw/device/sca/aes_serial.c
@@ -127,7 +127,7 @@ static void aes_key_mask_and_config(const uint8_t *key, size_t key_len) {
     key_shares.share0[i] = *((uint32_t *)key + i) ^ key_shares.share1[i];
   }
   // Provide random shares for unused key bits.
-  for (int i = key_len; i < kAesKeyLengthMax / 4; ++i) {
+  for (size_t i = key_len; i < kAesKeyLengthMax / 4; ++i) {
     key_shares.share1[i] = sca_non_linear_layer(sca_next_lfsr(1));
     key_shares.share0[i] = sca_non_linear_layer(sca_next_lfsr(1));
   }
